Actually replace deluge plugin with the new version.
